requirements.txt
django==4.1.2
djangorestframework==3.14.0
postgres==4.0
psycopg2==2.9.5
pillow==9.2.0
numpy==1.23.4
scipy==1.9.3